requirements.txt
ollama==0.3.3
click==8.1.7
rich==13.7.1
pyyaml==6.0.2
aider-chat==0.58.0

